Factored out product-create, product-purchase, and user-info apis (#55)
* sets up single api for producers to retail stream (may decide to locate elsewhere later, but lives with stream for moment).  todo set up the writing lambda behind the endpoint and a dynamodb for the schema against which endpoint needs to validate write requests

* added in lambda to back up writing endpoint.  valiudation to loop through all known schema.  still todo to store the schema somewhere else and not just list them in the code.

* changes to website to use new endpoint

* notes from exploring user info.  this has a few resources and roles that aren't being used.  of the ones that are, some policies within are not being used.  believe this should also be absorbed into single endpoint, but if it turns out that it needs to remain, should still trim off the code for unused resources.

* brought userInfoApi into the fold

* machete.  added version numbers in retail stream roles. removed unnecessary photo role.

* remove empty files

* removed removed services from deploy scripts.  remove CatoalogReaderRole from webapp

* Improve automated login on homepage.
Fixed issue with create product values having an extra character for a new product.
Removed deploy-web-app-plugin files which were used for reference, but not intended to be in project.

* Switched React router to use hashHistory instead of browserHistory, which allows the page to refresh and properly return the user to the correct location in the app.

* Removed unused devDependency.

* per Greg

* add in an install for the single api

* variable names fixed

* Removed unused devDependency.

* eslint 2

* cleanup per Erik
