+++ event-writer/schemas/product-create-schema.json
