+++ event-writer/schemas/product-purchase-schema.json
