+++ event-writer/schemas/retail-stream-schema-egress.json
