+++ event-writer/schemas/retail-stream-schema-ingress.json
