+++ package.json
@@ -8,24 +8,16 @@
     "lint": "./node_modules/.bin/eslint ./**/*.js ./**/*.jsx",
 
     "root:install":         "./build/0.env.sh && ./build/1.install.sh \"Installing Top-Level Project Dependencies\"            .",
-    "root:install:all":     "npm run root:install && npm run create:install && npm run purchase:install && npm run userInfo:install && npm run catalog:install:all && npm run photos:install:all && npm run web:install",
-    "root:deploy:all":      "npm run stream:deploy && npm run create:deploy && npm run purchase:deploy && npm run userInfo:deploy && npm run catalog:deploy:all && npm run photos:deploy:all && npm run web:deploy:all",
-    "root:remove:all":      "npm run web:remove:all && npm run photos:remove:all && npm run catalog:remove:all && npm run userInfo:remove && npm run purchase:remove && npm run create:remove && npm run stream:remove",
+    "root:install:all":     "npm run root:install && npm run writer:install && npm run catalog:install:all && npm run photos:install:all && npm run web:install",
+    "root:deploy:all":      "npm run stream:deploy && npm run writer:deploy && npm run catalog:deploy:all && npm run photos:deploy:all && npm run web:deploy:all",
+    "root:remove:all":      "npm run web:remove:all && npm run photos:remove:all && npm run catalog:remove:all && npm run writer:remove && npm run stream:remove",
+
+    "writer:install":       "./build/0.env.sh && ./build/1.install.sh \"Installing Event Writer Dependencies\"               ./event-writer",
+    "writer:deploy":        "./build/0.env.sh && ./build/2.sls.sh deploy \"Deploying Event Writer\"                          ./event-writer",
+    "writer:remove":        "./build/0.env.sh && ./build/2.sls.sh remove \"Removing Event Writer\"                           ./event-writer",
 
     "stream:deploy":        "./build/0.env.sh && ./build/2.sls.sh deploy \"Deploying Retail Stream\"                          ./retail-stream",
     "stream:remove":        "./build/0.env.sh && ./build/2.sls.sh remove \"Removing Retail Stream\"                           ./retail-stream",
-
-    "create:install":       "./build/0.env.sh && ./build/1.install.sh \"Installing Product Creation Dependencies\"            ./product-creation/api",
-    "create:deploy":        "./build/0.env.sh && ./build/2.sls.sh deploy \"Deploying Product Creation API\"                   ./product-creation/api",
-    "create:remove":        "./build/0.env.sh && ./build/2.sls.sh remove \"Removing Product Creation API\"                    ./product-creation/api",
-
-    "purchase:install":     "./build/0.env.sh && ./build/1.install.sh \"Installing Product Purchase Dependencies\"            ./product-purchase/api",
-    "purchase:deploy":      "./build/0.env.sh && ./build/2.sls.sh deploy \"Deploying Product Purchase API\"                   ./product-purchase/api",
-    "purchase:remove":      "./build/0.env.sh && ./build/2.sls.sh remove \"Removing Product Purchase API\"                    ./product-purchase/api",
-
-    "userInfo:install":     "./build/0.env.sh && ./build/1.install.sh \"Installing User Info Dependencies\"                   ./user-info/api",
-    "userInfo:deploy":      "./build/0.env.sh && ./build/2.sls.sh deploy \"Deploying User Info API\"                          ./user-info/api",
-    "userInfo:remove":      "./build/0.env.sh && ./build/2.sls.sh remove \"Removing User Info API\"                           ./user-info/api",
 
     "catalog:install:1":    "./build/0.env.sh && ./build/1.install.sh \"Installing Product Catalog Dependencies\"             ./product-catalog/builder",
     "catalog:install:2":    "./build/0.env.sh && ./build/1.install.sh \"Installing Product Catalog API\"                      ./product-catalog/api",
@@ -78,12 +70,6 @@
     "web:deploy:all":       "npm run web:build && npm run web:deploy",
     "web:remove:all":       "./build/0.env.sh && ./build/2.sls.sh remove \"Removing Web App\"                                 ./web",
 
-    "web:build:test": "pushd ./web/ && sls buildWebApp -s test",
-    "web:build:prod": "pushd ./web/ && sls buildWebApp -s prod",
-    "web:build:demo": "pushd ./web/ && sls buildWebApp -s demo",
-    "web:deploy:test": "pushd ./web/ && sls deploy -s test && popd",
-    "web:deploy:prod": "pushd ./web/ && sls deploy -s prod && popd",
-    "web:deploy:demo": "pushd ./web/ && sls deploy -s demo && popd",
     "web:dev:hot": "./node_modules/.bin/webpack-dev-server --config ./web/webpack.config.js --https --hot --progress --colors --watch --display-error-details --display-cached --content-base ./web/app/"
   },
   "repository": {
