--- product-creation/api/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "hello-retail-product-creation",
-  "version": "0.0.1",
-  "description": "Service to support events about product creation.",
-  "main": "catalogApi.js",
-  "scripts": {
-    "test": "echo \"Error: tests executed by root hello-retail package.json\" && exit 1"
-  },
-  "author": "Nordstrom",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/Nordstrom/hello-retail/issues"
-  },
-  "homepage": "https://github.com/Nordstrom/hello-retail/product-creation#readme",
-  "dependencies": {
-    "ajv": "^4.10.4"
-  }
-}
