--- product-creation/api/serverless.yml
@@ -1,111 +1,0 @@
-frameworkVersion: '>=1.0.0 <2.0.0'
-
-service: ${self:custom.helloRetail.name}-product-creation-api
-
-# TODO: Consider enforcing schema versions
-
-custom:
-  stage: ${opt:stage, self:provider.stage, self:custom.private.stage}
-  private: ${file(../../private.yml)}
-  helloRetail: ${file(../../helloRetail.yml)}
-  retailStreamVersion: 1
-  retailStream: ${file(../../retail-stream/serverless.yml)}
-  productCreationApiVersion: 1
-  productCreationApi: ${self:}
-  # Exports
-  exportProductCreationRestApi: ${self:custom.productCreationApi.service}:${self:custom.stage}:ProductCreationRestApi:${self:custom.productCreationApiVersion}
-
-provider:
-  name: aws
-  deploymentBucket: ${self:custom.helloRetail.deploymentBucket}
-  runtime: nodejs4.3
-  profile: ${self:custom.private.profile}
-  region: ${self:custom.private.region}
-
-functions:
-  productCreate:
-    role:
-      Fn::GetAtt: [ ProductCreationWriter, Arn ]
-    handler: productCreationApi.productCreate
-    environment:
-      STREAM_NAME:
-        Fn::ImportValue: ${self:custom.retailStream.custom.exportRetailStreamName}
-    events:
-      - http:
-          path: product-create
-          method: post
-          cors:
-            origins:
-              - '*'
-            headers:
-              - Content-Type
-              - X-Amz-Date
-              - Authorization
-              - X-Api-Key
-              - X-Amz-Security-Token
-              - X-Amz-User-Agent
-            allowCredentials: false
-
-resources:
-  Resources:
-    # Log Group
-    ProductCreationLogGroup:
-      Type: AWS::Logs::LogGroup
-      Properties:
-        LogGroupName: /aws/lambda/${self:service}-${self:custom.stage}-productCreate
-        RetentionInDays: 7
-    # User Information Roles
-    ProductCreationWriter: # role for the Product Create Lambdas
-      Type: AWS::IAM::Role
-      Properties:
-        Path: /
-        RoleName: ${self:custom.stage}ProductCreationWriter${self:custom.productCreationApiVersion}
-        AssumeRolePolicyDocument:
-          Version: '2012-10-17'
-          Statement:
-            - Effect: Allow
-              Action: sts:AssumeRole
-              Principal:
-                Service: lambda.amazonaws.com
-        ManagedPolicyArns:
-          - ${self:custom.private.teamPolicy}
-        Policies:
-          - PolicyName: CreateAndWriteToLogStream
-            PolicyDocument:
-              Version: '2012-10-17'
-              Statement:
-                - Effect: Allow
-                  Action:
-                    - 'logs:CreateLogStream'
-                  Resource:
-                    Fn::GetAtt: [ ProductCreationLogGroup, Arn ]
-                - Effect: Allow
-                  Action:
-                    - 'logs:PutLogEvents'
-                  Resource:
-                    - Fn::Join:
-                      - ':'
-                      - - Fn::GetAtt: [ ProductCreationLogGroup, Arn ]
-                        - '*'
-          - PolicyName: WriteToKinesis
-            PolicyDocument:
-              Version: '2012-10-17'
-              Statement:
-                - Effect: Allow
-                  Action:
-                    - 'kinesis:PutRecord'
-                  Resource:
-                    Fn::ImportValue: ${self:custom.retailStream.custom.exportRetailStreamArn}
-
-    # NOTE: This named gleaned from SLS generated CF template
-#    ApiGatewayMethodUserDashinfoPost:
-#      Properties:
-#        AuthorizationType: AWS_IAM
-
-  Outputs:
-    ProductCreationRestApi:
-      Description: API used to submit product creation events
-      Value:
-        Ref: ApiGatewayRestApi
-      Export:
-        Name: ${self:custom.productCreationApi.custom.exportProductCreationRestApi}
