--- product-purchase/api/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "hello-retail-product-purchase",
-  "version": "0.0.1",
-  "description": "Service to support putting the purchase of a single item onto the retail stream.",
-  "main": "catalogApi.js",
-  "scripts": {
-    "test": "echo \"Error: tests executed by root hello-retail package.json\" && exit 1"
-  },
-  "author": "Nordstrom",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/Nordstrom/hello-retail/issues"
-  },
-  "homepage": "https://github.com/Nordstrom/hello-retail/product-purchase#readme",
-  "dependencies": {
-    "ajv": "^4.10.4"
-  }
-}
