--- user-info/api/schemas/user-info-schema.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema#",
-  "self": {
-    "vendor": "com.nordstrom",
-    "name": "product/items",
-    "format": "jsonschema",
-    "version": "1-0-0"
-  },
-  "type": "array",
-  "items": {
-    "type": "object",
-    "properties": {
-      "schema": { "type": "string", "format": "url" },
-      "id":     { "type": "string" },
-      "name":   { "type": "string" },
-      "role":   { "type": "string" },
-      "phone":  { "type": "string" }
-    },
-    "required": [
-      "schema",
-      "id",
-      "brand",
-      "role",
-      "phone"
-    ],
-    "additionalProperties": false
-  },
-  "additionalProperties": false
-}
