+++ web/.serverless_plugins/web-app-build-plugin.js
@@ -126,7 +126,6 @@ class ServerlessPlugin {
       const provider = this.serverless.getProvider('aws')
       const importParts = outputName.split(':')
 
-      // TODO The fourth argument should be the version number, but currently, none of the userInfo, productCreation, or productPurchase APIs use versions (but should)
       if(importParts.length < 3) {
         reject(`Expected Fn::Import to match pattern: 'service:stage:outputName', but got '${outputName}' instead.`)
       }
