+++ web/components/products/product-detail-page.jsx
@@ -41,7 +41,7 @@ class ProductDetailPage extends Component {
   }
 
   purchaseProduct() {
-    this.props.awsLogin.makeApiRequest(config.ProductPurchaseAPI, 'POST', '/product-purchase/', {
+    this.props.awsLogin.makeApiRequest(config.EventWriterApi, 'POST', '/event-writer/', {
       schema: 'com.nordstrom/product/purchase/1-0-0',
       id: this.props.params.id,
       origin: `hello-retail/web-client-purchase-product/${this.props.awsLogin.state.profile.email}/${this.props.awsLogin.state.profile.name}`,
@@ -61,7 +61,7 @@ class ProductDetailPage extends Component {
       })
 
     this.setState({
-      buyMessage: 'On order.',
+      buyMessage: 'Submitting order.',
     })
   }
